added Reentrent, all works

+++ src/Server/ServerStrategySolveSearchProblem.java
@@ -1,22 +1,23 @@
 package Server;
 
-import algorithms.mazeGenerators.IMazeGenerator;
 import algorithms.mazeGenerators.Maze;
-import algorithms.mazeGenerators.Position;
 import algorithms.search.*;
 import sun.awt.Mutex;
 
 import java.io.*;
+import java.util.concurrent.locks.ReentrantLock;
 
 public class ServerStrategySolveSearchProblem implements IServerStrategy {
 
-    private Mutex mutex;
+    private ReentrantLock mutex;
+
 
     public ServerStrategySolveSearchProblem(){
-        mutex = new Mutex();
+        mutex = new ReentrantLock();
     }
 
     private static String tempDirectoryPath = System.getProperty("java.io.tmpdir");
+
     @Override
     public void serverStrategy(InputStream inFromClient, OutputStream outToClient) {
 
@@ -28,35 +29,35 @@ public class ServerStrategySolveSearchProblem implements IServerStrategy {
             try {
                 Maze maze = (Maze) fromClient.readObject();
                 String tempPath = tempDirectoryPath + "maze" + maze.toString().hashCode();
-                System.out.println(tempPath);
+
                 mutex.lock();
                 File f = new File(tempPath);
                 if(f.exists()){ // the file exists, we don't need to solve again. only take what exists.
-                    System.out.println("####### File Exists ######## ");
                     FileInputStream fin = new FileInputStream(tempPath);
                     ObjectInputStream oin = new ObjectInputStream(fin);
                     solution = (Solution) oin.readObject();
-                    System.out.println("#######solution is:#######\n" + solution+ "\n\n####################");
                     fin.close();
                     oin.close();
+                    mutex.unlock();
                 }
                 else{
-                    System.out.println("####### File NOT Exists ######## ");
-
+                    mutex.unlock();
                     SearchableMaze searchableMaze = new SearchableMaze(maze);
                     solution = Configurations.getAlgorithms_solveAlgorithm().solve(searchableMaze);
-                    System.out.println("#######solution is:#######\n" + solution+ "\n\n####################");
-                    f.createNewFile();
-                    FileOutputStream fout = new FileOutputStream(tempPath);
-                    fout.flush();
-                    ObjectOutputStream oout = new ObjectOutputStream(fout);
-                    oout.flush();
-                    oout.writeObject(solution);
-                    oout.flush();
-                    fout.close();
-                    oout.close();
+                    mutex.lock();
+                    if(!f.exists()) {
+                        f.createNewFile();
+                        FileOutputStream fout = new FileOutputStream(tempPath);
+                        fout.flush();
+                        ObjectOutputStream oout = new ObjectOutputStream(fout);
+                        oout.flush();
+                        oout.writeObject(solution);
+                        oout.flush();
+                        fout.close();
+                        oout.close();
+                    }
+                    mutex.unlock();
                 }
-                mutex.unlock();
                 toClient.writeObject(solution);
                 toClient.flush();
             }
